reorganize users_handlers & move validation init to separate file
