+++ internal/http/handlers/auth/login.go
@@ -1,0 +1,81 @@
+package auth
+
+import (
+	"net/http"
+	"os"
+	"time"
+
+	"github.com/Sofassssssss/Board-game-tracking/internal/config"
+	gormrepo "github.com/Sofassssssss/Board-game-tracking/internal/repo/gorm"
+	"github.com/Sofassssssss/Board-game-tracking/internal/repo/gorm/models"
+	"github.com/gin-gonic/gin"
+	"github.com/golang-jwt/jwt/v5"
+	"golang.org/x/crypto/bcrypt"
+)
+
+func Login(c *gin.Context) {
+	// Get the email and password of req body
+	var body struct {
+		Email    string
+		Password string
+	}
+
+	if c.Bind(&body) != nil {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"error": "Failed to read body",
+		})
+
+		return
+	}
+
+	// Look up requested user
+	var user models.User
+	gormrepo.DB.First(&user, "email = ?", body.Email)
+
+	if user.ID == 0 {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"error": "Invalid email",
+		})
+		return
+	}
+
+	// Compare sent in password with saved user password hash
+	err := bcrypt.CompareHashAndPassword([]byte(user.HashedPassword), []byte(body.Password))
+
+	if err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"error": "Invalid password",
+		})
+		return
+
+	}
+
+	// Generate a jwt token
+	token := jwt.NewWithClaims(jwt.SigningMethodHS256, jwt.MapClaims{
+		"sub": user.ID,
+		"exp": time.Now().Add(time.Hour * 24 * 30).Unix(),
+	})
+
+	tokenString, err := token.SignedString([]byte(os.Getenv("SECRET")))
+
+	if err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"error": "Failed to create token",
+		})
+		return
+	}
+
+	// Send it back
+	c.SetSameSite(http.SameSiteLaxMode)
+	c.SetCookie(
+		"Authorization",
+		tokenString,
+		int(config.TokenMaxAge.Seconds()),
+		"",
+		"",
+		false,
+		true)
+
+	c.JSON(http.StatusOK, gin.H{})
+
+}
