+++ internal/http/router/router.go
@@ -1,13 +1,13 @@
 package router
 
 import (
-	"github.com/Sofassssssss/Board-game-tracking/internal/http/handlers"
+	"github.com/Sofassssssss/Board-game-tracking/internal/http/handlers/auth"
 	"github.com/Sofassssssss/Board-game-tracking/internal/http/middleware"
 	"github.com/gin-gonic/gin"
 )
 
 func SetupRoutes(r *gin.Engine) {
-	r.POST("/signup", handlers.Signup)
-	r.POST("/login", handlers.Login)
-	r.GET("/validate", middleware.RequireAuth, handlers.Validate)
+	r.POST("/signup", auth.Signup)
+	r.POST("/login", auth.Login)
+	r.GET("/validate", middleware.RequireAuth, auth.Validate)
 }
