+++ internal/http/validators/init.go
@@ -1,0 +1,12 @@
+package validators
+
+import (
+	"github.com/go-playground/validator/v10"
+)
+
+var Validate *validator.Validate
+
+func Init() {
+	Validate = validator.New()
+	_ = Validate.RegisterValidation("password", PasswordValidator)
+}
